Edit metricThreshold api to controller&service
