+++ backend/main.py
@@ -3,26 +3,20 @@
 #########
 # internal imports
 from db import db
-from models.infrastructureConfigModel import InfrastructureConfig
 from models.serverLogsModel import ServerLogs
-from models.webAppLogsModel import WebAppLogs
-from models.databaseLogsModel import DatabaseLogs
-from models.serverHealthStatusThresholdsModel import ServerHealthStatusThresholds
 from infrastructureConfig.infrastructureConfigController import infrastructureConfigBp
 from serverLogs.serverLogsController import serverLogsBp
 from webAppLogs.webAppLogsController import webAppLogsBp
 from databaseLogs.databaseLogsController import databaseLogsBp
+from metricThreshold.metricThresholdController import metricThresholdBp
 
 # external imports
 from dotenv import load_dotenv
 import os
-import json
-from flask import Flask, jsonify, request
-from flask_sqlalchemy import SQLAlchemy
+from flask import Flask
 from flask_socketio import SocketIO, emit
-from sqlalchemy import text, desc, func, and_ #func is a module provided by SQLAlchemy that allows usage of SQL functions in queries. and_ is a logical operator provided by SQLAlchemy that constructs an SQL AND clause.
+from sqlalchemy import text, func, and_ #func is a module provided by SQLAlchemy that allows usage of SQL functions in queries. and_ is a logical operator provided by SQLAlchemy that constructs an SQL AND clause.
 from flask_cors import CORS
-import pyodbc
 
 
 #############################
@@ -44,6 +38,7 @@ app.register_blueprint(infrastructureConfigBp)
 app.register_blueprint(serverLogsBp)
 app.register_blueprint(webAppLogsBp)
 app.register_blueprint(databaseLogsBp)
+app.register_blueprint(metricThresholdBp)
 
 ################
 # DEFAULT ROUTES
@@ -74,71 +69,7 @@ def check_db_connection():
         return False
 
 
-# Check the health status of the servers for server_logs table
-@app.route('/health_status', methods=['GET'])
-def get_health_status():
-    try:
-        # Fetch the latest server logs
-        latest_server_logs_response = get_latest_server_logs()
-        response, status_code = get_latest_server_logs()
-
-        # Check if the response is successful
-        if status_code == 200:
-            # Extract the data from the response
-            json_data = response.json
-
-            latest_server_logs = json_data['data']['latest_server_logs']
-
-            # Fetch health status thresholds from the database
-            thresholds = ServerHealthStatusThresholds.query.all()
-            threshold_dict = {threshold.Metric: (threshold.CriticalThreshold, threshold.BadThreshold, threshold.WarningThreshold) for threshold in thresholds}
-
-            # List to store health status data
-            health_status_data = []
-
-            # Iterate through each server log
-            for log in latest_server_logs:
-                server_name = log['ServerName']
-                health_status = {}
-
-                # Determine health status for each metric
-                for metric in log.keys():
-                    if metric in threshold_dict:
-                        value = log[metric]
-                        critical_threshold, bad_threshold, warning_threshold = threshold_dict[metric]
-
-                        if metric == 'Availability' or metric == 'NetworkAvailability':
-                            if value == 0:
-                                health_status[metric] = 'Critical'
-                            else:
-                                health_status[metric] = 'Healthy'
-                        else:
-                            if value >= critical_threshold:
-                                health_status[metric] = 'Critical'
-                            elif value >= bad_threshold:
-                                health_status[metric] = 'Bad'
-                            elif value >= warning_threshold:
-                                health_status[metric] = 'Warning'
-                            else:
-                                health_status[metric] = 'Healthy'
-
-                # Append server name and its health status to the list
-                health_status_data.append({'ServerName': server_name, 'HealthStatus': health_status})
-
-            if health_status_data:
-                return jsonify({"code": 200, "data": health_status_data}), 200
-            else:
-                return jsonify({"code": 404, "message": "No server logs available."}), 404
-        else:
-            # Return the response from get_latest_server_logs() directly
-            return latest_server_logs_response[0]
-    except Exception as e:
-        # Handle any exceptions and return an error response
-        return jsonify({"code": 500, "message": f"An error occurred: {str(e)}"}), 500
-
-
 #---SocketIO---#
-
 @socketio.on('serverlog', namespace='/latestlogs')
 def handle_server_logs_connect():
     try:
