+++ backend/metricThreshold/metricThresholdController.py
@@ -1,0 +1,16 @@
+from flask import jsonify, Blueprint
+from metricThreshold.metricThresholdService import metricThresholdClass
+
+# initialise blueprint
+metricThresholdBp = Blueprint('metricThreshold', __name__, url_prefix='/metricthreshold')
+
+# initialise route and service to call
+# Check the health status of the servers for server_logs table
+@metricThresholdBp.route('/health_status', methods=['GET'])
+def get_health_status():
+    try:
+        response = metricThresholdClass.get_health_status()
+        return response
+    except Exception as e:
+        return jsonify({'error': str(e)}), 500
+
